Comment out logic that prevents Layout and SideNav from being server components.
