+++ frontend/src/app/components/layouts/sideNav/SideNav.module.scss
@@ -1,4 +1,4 @@
-@import "../variables.module.scss";
+@import "../../../variables.module.scss";
 
 .sideNavContainer {
   display: flex;
